contents/base/servers/libraries/drivers/opensearch/requirements.txt
requests
boto3
requests-aws4auth
opensearch-py
python-dotenv
httpx